requirements.txt
# Deep Learning and Computer Vision
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0

# Google Earth Engine
earthengine-api==0.1.374
google-auth==2.23.3
google-auth-oauthlib==1.1.0

# Web Framework
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7

# Data Handling
pandas==2.0.3
requests==2.31.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
folium==0.14.0
